requirements.txt
flask==2.3.3
paddlepaddle==2.6.2
paddleocr==2.7.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
werkzeug==2.3.7